Added logic for dice + Reimplemented Grid
Added logic for player to move based on dice value.
Also changed the code that initializes the grid and made it such that the tile position alternates just like in a real snakes and ladders game. 0-9 go from left to right then 10 is above 9 and 10-19 go from right to left and so on.