Fixed bad code + Added Sharks and Turtles
Fixed bad code that resulted in objects being created within class resulting in duplicate nodes for each objects. This seriously brought the frame rate down. Also updated tileArray type to be a structure rather than a Sprite Node as there was no use of it other than defining boundaries. This resulted in significant increase in efficiency and decrease in redundancy.

Also added class for Sharks and Turtles and programmed logic for when the player lands on a tile with them.

Fixed the bug where multiple touch inputs would seriously mess up the game. The game now disables touch input and enables it only after all the actions have been completed. An extra delay of 100ms is added when enabling touch inputs.