Added Player 2 + Removed hardcoded size and position from classes.
Added player 2 to be used as player 2 or computer. Also fixed positioning issues which resulted from player 2 being added. That uncovered bad coding that had sizes and positions hardcoded in the classes. Classes are meant to be abstract and size should be defined by the code creating the object. Fixed that for all the classes.

Next step is to create different scenes, main menu, pause menu, etc.