Added Main Menu and Game Menu using storyboards.
A temporary main menu and game menu have been added using storyboards that link to the main game. The layout has been decided but can change. The graphics for the menus will be done at the end phases along with the graphics for the main game and will be concurrent with the theme of the game.