Added camera node + experimented with perspective shift.
Added camera node on the main game in case animation with the camera are required. Experimented with close camera panning for when the player lands on a shark or title and also perspective shift however they disrupt the flow of the game.
TODO - Try to add panning and perspective for the winning move and check if it makes sense adding an animation there.