Modified game to single view application + Major refactoring.
Modified game to have a single view managed by one view controller. This helps to perform better animations when moving through the game by changing scenes rather than performing segues. Main menu and game menu are now SpriteKit scenes and are ready to be textured.

Major refactoring done to fix many memory leaks and to make the code readable and efficient. Deleted the segues and now the game does transitions programmatically as and when needed. Fixed redundancies in the code. Removed the camera node as it was of no use.